Add files via upload
this is the final size and look of the project, after many iterations, we found that this is the best configuration of the sytem. the files can be changed to what you need , but sizing of thngs like the down ramp, ball Corrall, and ball ramp must stay the exact for the connections of it . please note that when creating this design the solidworks function (shell) was used , at your own discretion the sizes of walls and how they fit together is up to you, we found the having all top pieces the proper lencgth as in the file, and then change tp the side wall is the best way to keep track.